move defaultValue update to separate plugin
